backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.5.2",
    "boto3==1.24.56",
    "certifi==2022.12.7",
    "charset-normalizer==2.1.1",
    "coreapi==2.3.3",
    "coreschema==0.0.4",
    "daphne==4.0.0",
    "django==4.1.4",
    "django-cloudinary-storage==0.3.0",
    "django-cors-headers==3.13.0",
    "django-jazzmin==2.6.1",
    "django-storages==1.13.2",
    "djangorestframework==3.14.0",
    "drf-yasg==1.21.4",
    "idna==3.4",
    "inflection==0.5.1",
    "itypes==1.2.0",
    "jinja2==3.1.2",
    "markupsafe==2.1.1",
    "packaging==21.3",
    "pyjwt==2.6.0",
    "pyparsing==3.0.9",
    "python-dotenv==0.21.0",
    "pytz==2022.6",
    "requests==2.28.1",
    "sentry-sdk==2.18.0",
    "sqlparse==0.4.3",
    "twisted[http2,tls]>=24.11.0",
    "tzdata==2022.7",
    "uritemplate==4.1.1",
    "urllib3==1.26.13",
    "whitenoise==6.8.2",
]
